Add files via upload
The only change/file that is inside of the "project" was this python script. I didn't exactly believe I would do well, but hey. It's over with, and I would like the people of GitHub judge me on how I wrote. Don't expect many coding comments because I didn't think to do that considering it was for myself.